+++ IPA Chat/IPA_ChatApp.swift
@@ -9,19 +9,9 @@ import SwiftUI
 
 @main
 struct IPA_ChatApp: App {
-    @StateObject var voiceEngine: VoiceEngine = VoiceEngine()
-
     var body: some Scene {
         WindowGroup {
             ContentView()
-                .environmentObject(voiceEngine)
-                .onAppear {
-                    voiceEngine.load()
-                }
-                .onDisappear {
-                    voiceEngine.save()
-                }
-
         }
     }
 }
