+++ IPA Chat/Module/ContentView1.swift
@@ -1,244 +1,6 @@
 import SwiftUI
-import AVFoundation
-import SQLite
 
-struct VoiceWrapper: Hashable {
-    let id: UUID
-    let voice: AVSpeechSynthesisVoice
-    
-    init(voice: AVSpeechSynthesisVoice) {
-        self.id = UUID()
-        self.voice = voice
-    }
-}
-
-func voicesByLanguage() -> [String: [VoiceWrapper]] {
-    let voices = AVSpeechSynthesisVoice.speechVoices()
-    var voiceDict = [String: [VoiceWrapper]]()
-    
-    for voice in voices {
-        let wrapper = VoiceWrapper(voice: voice)
-        let languageCode = voice.language
-        if voiceDict[languageCode] != nil {
-            voiceDict[languageCode]!.append(wrapper)
-        } else {
-            voiceDict[languageCode] = [wrapper]
-        }
-    }
-    
-    return voiceDict
-}
-
-
-// Manually implement Codable to ignore the `color` property
-enum CodingKeys: String, CodingKey {
-    case id, symbol, ipaNotation, type, x, y
-}
-
-enum PhonemeType: Codable {
-    case vowel, consonant, diphthong, stress, pause, semivowel, nasal
-}
-
-struct Phoneme: Identifiable, Codable {
-    let id: UUID  // Make this immutable
-    let symbol: String
-    let ipaNotation: String
-    let type: PhonemeType
-    var x: CGFloat
-    var y: CGFloat
-    
-    init(id: UUID? = nil, symbol: String, ipaNotation: String, type: PhonemeType, x: CGFloat = 0, y: CGFloat = 0) {
-        self.id = id ?? UUID()
-        self.symbol = symbol
-        self.ipaNotation = ipaNotation
-        self.type = type
-        self.x = x
-        self.y = y
-    }
-    
-    var color: Color {
-        switch type {
-        case .vowel: return Color.red
-        case .consonant: return Color.blue
-        case .diphthong: return Color.green
-        case .stress: return Color.purple
-        case .semivowel: return Color.brown
-        case .nasal: return Color.gray
-        case .pause: return Color.pink
-        }
-    }
-    
-    enum CodingKeys: String, CodingKey {
-        case id, symbol, ipaNotation, type, x, y
-    }
-}
-
-
-// The observable object
-class ObservablePhoneme: ObservableObject {
-    @Published var phoneme: Phoneme
-    
-    init(phoneme: Phoneme) {
-        self.phoneme = phoneme
-    }
-}
-
-class AudioManager: ObservableObject {
-    private var synthesizer = AVSpeechSynthesizer()
-    public var phonemeString = NSMutableAttributedString(string: "")
-    @Published var currentPhonemeSequence: String = ""
-    @Published var isBabbleModeOn: Bool = false
-    @Published var shouldSpeakFullUtterance: Bool = true  // Add this for the setting
-    @Published var selectedVoice: AVSpeechSynthesisVoice? = AVSpeechSynthesisVoice(language: "en-GB")
-    var db: Connection!
-    
-    init () {
-        // Initialize the audio session to play sound even in silent mode
-        do {
-            try AVAudioSession.sharedInstance().setCategory(.playback, mode: .default, options: [])
-            try AVAudioSession.sharedInstance().setActive(true)
-        } catch {
-            print("Failed to set up audio session: \(error)")
-        }
-        connectToDB()
-    }
-    
-    func connectToDB() {
-        if let sourceURL = Bundle.main.url(forResource: "ipa-dict", withExtension: "db") {
-            do {
-                db = try Connection(sourceURL.path)
-            } catch {
-                print("Connection failed: \(error)")
-            }
-        }
-    }
-    
-    func getPhonemeForString(selectedLanguage: String, searchString: String) -> String? {
-        guard let db = self.db else {
-            print("Database not initialized.")
-            return nil
-        }
-        
-        let tableName = selectedLanguage == "French" ? "fr_FR" : "en_UK"
-        let ipa = Expression<String>("ipa")
-        let string = Expression<String>("string")
-        let table = Table(tableName)
-        
-        do {
-            let query = table.select(ipa).filter(string == searchString)
-            for row in try db.prepare(query) {
-                let ipaCol = row[ipa]
-                print("Found IPA: \(ipaCol)")
-                return ipaCol
-            }
-        } catch {
-            print("Select failed: \(error)")
-        }
-        
-        print("No IPA found")
-        return nil
-    }
-    
-    func appendPhoneme(_ phoneme: Phoneme) {
-        let str = "\(phoneme.symbol)" // space for separation
-        let newPhonemeString = NSMutableAttributedString(string: str)
-        let range = NSRange(location: 0, length: str.count)
-        newPhonemeString.addAttribute(.init(AVSpeechSynthesisIPANotationAttribute), value: phoneme.ipaNotation, range: range)
-        
-        // Speak just the new phoneme if Babble Mode is on
-        if isBabbleModeOn {
-            let utterance = AVSpeechUtterance(attributedString: newPhonemeString)
-            utterance.voice = selectedVoice
-            synthesizer.speak(utterance)
-            return
-        }
-        
-        // Otherwise, append the new phoneme to the sequence and optionally speak the full sequence
-        phonemeString.append(newPhonemeString)
-        currentPhonemeSequence += "\(phoneme.symbol)"
-        
-        if shouldSpeakFullUtterance {
-            let utterance = AVSpeechUtterance(attributedString: phonemeString)
-            utterance.voice = selectedVoice
-            synthesizer.speak(utterance)
-        }
-    }
-    
-    func speakCurrentSequence() {
-        // Use the phonemeString with IPA notation to speak the sequence
-        let utterance = AVSpeechUtterance(attributedString: phonemeString)
-        utterance.voice = selectedVoice
-        if let voice = utterance.voice {
-            print("Voice Name: \(voice.name)")
-            print("Voice Language: \(voice.language)")
-            print("Voice Quality: \(voice.quality.rawValue)")
-        } else {
-            print("No voice set for the utterance.")
-        }
-        
-        synthesizer.speak(utterance)
-    }
-    
-    func clearSequence() {
-        currentPhonemeSequence = ""
-        phonemeString = NSMutableAttributedString(string: "")
-    }
-    
-    func toggleButton() {
-        // Perform the toggle logic here.
-        // For example, if you have a boolean property called `isBabbleModeOn`
-        isBabbleModeOn.toggle()
-    }
-}
-
-struct SettingsView: SwiftUI.View {
-    @ObservedObject var audioManager: AudioManager
-    @SwiftUI.Binding var selectedLanguage: String
-    var languages = ["English-GB", "French"]
-    var groupedVoices: [String: [VoiceWrapper]] = voicesByLanguage()
-    @SwiftUI.Binding var phonemes: [Phoneme]
-    
-    var body: some SwiftUI.View {
-        Form {
-            Toggle("Speak utterance as selected", isOn: $audioManager.shouldSpeakFullUtterance)
-            
-            Picker("Language", selection: $selectedLanguage) {
-                ForEach(languages, id: \.self) {
-                    Text($0)
-                }
-            }
-            
-            Picker("Select Voice", selection: $audioManager.selectedVoice) {
-                ForEach(groupedVoices.keys.sorted(), id: \.self) { language in
-                    Section(header: Text(language)) {
-                        ForEach(groupedVoices[language]!, id: \.self) { wrapper in
-                            Text(wrapper.voice.name).tag(wrapper.voice as AVSpeechSynthesisVoice?)
-                        }
-                    }
-                }
-            }
-            
-            Section(header: Text("Reorder Phonemes")) {
-                List {
-                    ForEach(phonemes) { phoneme in
-                        Text(phoneme.symbol)
-                    }
-                    .onMove(perform: move)
-                }
-            }
-        }.toolbar {
-            EditButton()
-        }
-    }
-    
-    func move(from source: IndexSet, to destination: Int) {
-        // Move elements around
-        phonemes.move(fromOffsets: source, toOffset: destination)
-    }
-    
-}
-
-struct ContentView: SwiftUI.View {
+struct ContentView: View {
     @Environment(\.horizontalSizeClass) var horizontalSizeClass
     @ObservedObject var audioManager = AudioManager()
     @State var isEditMode: Bool = false
@@ -360,7 +122,7 @@ struct ContentView: SwiftUI.View {
     }
     
     
-    var body: some SwiftUI.View {
+    var body: some View {
         GeometryReader { geometry in
             NavigationView {
                 VStack {
