--- IPA Chat/Observables/AvailableVoices.swift
@@ -1,70 +1,0 @@
-//
-//  AvailableVoices.swift
-//  Echo
-//
-//  Created by Gavin Henderson on 03/10/2023.
-//
-
-import Foundation
-import AVKit
-
-class AvailableVoices: ObservableObject {
-    @Published var voices: [AVSpeechSynthesisVoice] = []
-    @Published var voicesByLang: [String: [AVSpeechSynthesisVoice]]
-    
-    init() {
-        voices = AVSpeechSynthesisVoice.speechVoices()
-        voicesByLang = [:]
-        for voice in voices {
-            var currentList = voicesByLang[voice.language] ?? []
-            currentList.append(voice)
-            voicesByLang[voice.language] = currentList
-        }
-    }
-    
-    /***
-        Sorts all the languages availble by the following criteria
-     
-        * Push users language and region to the top
-        * Push users language to the top
-        * Sort alphabetically (by display name)
-     */
-    func sortedKeys() -> [String] {
-        let currentLocale: String = Locale.current.language.languageCode?.identifier ?? "en"
-        let currentIdentifier: String = Locale.current.identifier(.bcp47)
-        
-        return Array(voicesByLang.keys).sorted(by: {
-            let zeroLocale = Locale(identifier: $0).language.languageCode?.identifier ?? "en"
-            let oneLocale = Locale(identifier: $1).language.languageCode?.identifier ?? "en"
-            let zeroFullLanguage = getLanguageAndRegion($0)
-            let oneFullLanguage = getLanguageAndRegion($1)
-            
-            if $0 == currentIdentifier {
-                return true
-            }
-            
-            if $1 == currentIdentifier {
-                return false
-            }
-            
-            if zeroLocale == currentLocale && oneLocale == currentLocale {
-                return zeroFullLanguage < oneFullLanguage
-            }
-            
-            if zeroLocale == currentLocale {
-                return true
-            }
-            
-            if oneLocale == currentLocale {
-                return false
-            }
-            
-            return zeroFullLanguage < oneFullLanguage
-
-        })
-    }
-    
-    func voicesForLang(_ lang: String) -> [AVSpeechSynthesisVoice] {
-        return self.voicesByLang[lang] ?? []
-    }
-}
